Finish Dish Detailed Activity

+++ ElAkil/app/src/main/java/com/example/elakil/presentation/favorites/FavoritesPresenter.java
@@ -1,8 +1,16 @@
 package com.example.elakil.presentation.favorites;
 
+import android.os.Handler;
+import android.os.Looper;
+
+
 import com.example.elakil.data.MealsRepository;
 import com.example.elakil.model.Meal;
 import com.example.elakil.utils.SharedPreferencesUtils;
+
+import java.util.List;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
 public class FavoritesPresenter implements FavoritesContract.Presenter{
 
@@ -12,21 +20,26 @@ public class FavoritesPresenter implements FavoritesContract.Presenter{
 
     private SharedPreferencesUtils sharedPreferencesUtils;
 
+    private ExecutorService executorService ;
+    private Handler mainHandler ;
+
     public FavoritesPresenter(FavoritesContract.View view, MealsRepository repository, SharedPreferencesUtils sharedPreferencesUtils) {
-        this.view = view;
+        this.view = view            ;
         this.repository = repository;
         this.sharedPreferencesUtils = sharedPreferencesUtils;
+        this.executorService = Executors.newSingleThreadExecutor();
+        this.mainHandler = new Handler(Looper.getMainLooper());
     }
 
     @Override
     public void loadFavoriteMeals() {
         if (sharedPreferencesUtils.isGuestMode()){
-            view.showGuestMessage();
-        }
-        else {
-            repository.getFavoriteMeals().observeForever(meals -> {
+            mainHandler.post(() -> view.showGuestMessage());
+        } else {
+            executorService.execute(() -> {
+                List<Meal> meals = repository.getFavoriteMeals().getValue();
                 if (meals != null){
-                    view.showFavoritesMeals(meals);
+                    mainHandler.post(() -> view.showFavoritesMeals(meals));
                 }
             });
         }
@@ -34,10 +47,20 @@ public class FavoritesPresenter implements FavoritesContract.Presenter{
 
     @Override
     public void removeFromFavorites(Meal meal) {
-        meal.setFavorite(false);
-        repository.deleteMeals(meal , success -> {
-            loadFavoriteMeals();
-        });
+        if (meal != null){
+            executorService.execute(() -> {
+                meal.setFavorite(false);
+                repository.deleteMeals(meal , success -> {
+                    mainHandler.post(() -> {
+                        if (success){
+
+                        } else {
+
+                        }
+                    });
+                });
+            });
+        }
 
     }
